+++ Sources/Internal/Render/RHI/Common/MCPP/expand.cpp
@@ -2191,7 +2191,7 @@ static char *   expand_prestd(
         case NUM:                           /* Number token         */
         case OPE:                           /* Operator or punct.   */
         case NAM:                           /* Identifier           */
-            len = strlen( mp);
+            len = static_cast<int>(strlen( mp));
             mp += len;
             break;
         case SEP:                           /* Special character    */
